requirements.txt
numpy == 1.21.1
pandas == 1.3.0
uvicorn == 0.20.0
scipy == 1.7.1
implicit == 0.6.1
fastapi == 0.88.0
python-dotenv == 1.0.0